requirements-cpu.txt
absl-py==1.4.0
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.6
colorama==0.4.6
comtypes==1.2.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
dxcam==0.0.5
executing==1.2.0
filelock==3.9.0
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
idna==3.4
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
keyboard==0.13.5
kiwisolver==1.4.4
loguru==0.7.0
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.2.1
mypy-extensions==1.0.0
networkx==3.0
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
thop==0.1.1.post2209072238
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.3.6
win32-setctime==1.1.0
